python/requirements.txt
web3>=6.15.1
aiohttp>=3.9.1
python-dotenv>=1.0.0
